start: move start methods to start-remote.js so that it can be run independently
